Cleanup

+++ ssh/install.fish
@@ -1,5 +1,5 @@
 #!/usr/bin/env fish
-test -L ~/.ssh/config ||
+if test -e ~/.ssh/config
 	mv ~/.ssh/config ~/.ssh/config.local
 	and ln -s "$DOTFILES"/ssh/config ~/.ssh/config
 end
